Final ops layer cleanup: misplaced logic and duplicate utilities
Completes the operations layer project by moving remaining misplaced
logic to proper layers and eliminating duplicated utilities:

- Move classify_aliases (pure expression/key classification) from
  alias.rs to expr/alias.rs where the other pure alias logic lives
- Move filter_by_roots (root spec parsing, lock entry filtering) from
  apply.rs to ops/apply.rs
- Extract WorklistEntry construction and unique-content deduplication
  from worklist.rs to new ops/worklist.rs module
- Consolidate format_count: make ceremony::format_count generic over
  Display, delete duplicate format_number from coverage.rs and facts.rs,
  move tests to ceremony.rs, remove unnecessary as usize casts

+++ src/import_facts.rs
@@ -66,14 +66,14 @@ pub fn run(db: &mut Db, allow_archived: bool, verbose: bool) -> Result<()> {
     let stats = &state.stats;
     println!(
         "Processed {} lines: {} facts imported, {} skipped (stale), {} skipped (reserved), {} skipped (archived), {} skipped (type mismatch), {} objects created, {} facts promoted",
-        ceremony::format_count(stats.lines_processed as usize),
-        ceremony::format_count(stats.facts_imported as usize),
-        ceremony::format_count(stats.skipped_stale as usize),
-        ceremony::format_count(stats.skipped_reserved as usize),
-        ceremony::format_count(stats.skipped_archived as usize),
-        ceremony::format_count(stats.skipped_type_mismatch as usize),
-        ceremony::format_count(stats.objects_created as usize),
-        ceremony::format_count(stats.facts_promoted as usize)
+        ceremony::format_count(stats.lines_processed),
+        ceremony::format_count(stats.facts_imported),
+        ceremony::format_count(stats.skipped_stale),
+        ceremony::format_count(stats.skipped_reserved),
+        ceremony::format_count(stats.skipped_archived),
+        ceremony::format_count(stats.skipped_type_mismatch),
+        ceremony::format_count(stats.objects_created),
+        ceremony::format_count(stats.facts_promoted)
     );
 
     // Update query planner statistics after bulk changes
